Split iOS AI chat types into focused modules
