+++ apps/ios/Flashcards/Flashcards/AI/AIChatAccessSupport.swift
@@ -1,0 +1,28 @@
+import Foundation
+
+enum AIChatAccessState: Equatable {
+    case consentRequired
+    case ready
+}
+
+func hasAIChatExternalProviderConsent(userDefaults: UserDefaults) -> Bool {
+    userDefaults.bool(forKey: aiChatExternalProviderConsentUserDefaultsKey)
+}
+
+func grantAIChatExternalProviderConsent(userDefaults: UserDefaults) {
+    userDefaults.set(true, forKey: aiChatExternalProviderConsentUserDefaultsKey)
+}
+
+func aiChatAccessState(
+    hasExternalProviderConsent: Bool
+) -> AIChatAccessState {
+    guard hasExternalProviderConsent else {
+        return .consentRequired
+    }
+
+    return .ready
+}
+
+func isGuestAiLimitCode(_ code: String?) -> Bool {
+    code == "GUEST_AI_LIMIT_REACHED"
+}
