+++ apps/ios/Flashcards/Flashcards/AI/AIChatContentModels.swift
@@ -1,76 +1,5 @@
 import Foundation
 
-let aiChatDefaultModelId: String = "gpt-5.4"
-let aiChatDefaultModelLabel: String = "GPT-5.4"
-let aiChatDefaultProviderLabel: String = "OpenAI"
-let aiChatDefaultReasoningEffort: String = "medium"
-let aiChatDefaultReasoningLabel: String = "Medium"
-let aiChatClientPlatform: String = "ios"
-let aiChatCreateCardDraftPrompt: String = "Help me create a card."
-let aiChatOptimisticAssistantStatusText: String = "Looking through your cards..."
-let aiChatExternalProviderConsentUserDefaultsKey: String = "ai-chat-external-provider-consent"
-let aiChatExternalProviderConsentRequiredMessage: String = "Review AI data use and accept it on this device before using AI features."
-let aiChatAccuracyWarningText: String = "AI responses can be inaccurate or incomplete. Review important results before relying on them."
-let aiChatGuestQuotaReachedMessage: String = "Your free guest AI limit for this month is used up. Create an account or log in to keep using AI."
-let aiChatGuestQuotaButtonTitle: String = "Create account or Log in"
-let aiChatMaximumAttachmentBytes: Int = 20 * 1024 * 1024
-let aiChatLocalSessionStalenessThreshold: TimeInterval = 6 * 60 * 60
-let aiChatSupportedFileExtensions: Set<String> = [
-    "pdf",
-    "txt",
-    "csv",
-    "json",
-    "xml",
-    "xlsx",
-    "xls",
-    "md",
-    "html",
-    "py",
-    "js",
-    "ts",
-    "yaml",
-    "yml",
-    "sql",
-    "log",
-    "docx",
-]
-let aiChatToolNames: Set<String> = [
-    "sql",
-]
-let aiChatExternalProviderDisclosureItems: [String] = [
-    "Typed prompts and card-derived context needed for your request can be sent to OpenAI.",
-    "Uploaded files and images can be uploaded to OpenAI for AI processing.",
-    "Dictated audio and transcription requests can be sent to OpenAI for speech processing.",
-    "Technical diagnostics about failed or slow AI requests can be sent to help debug the hosted AI service.",
-]
-
-enum AIChatAccessState: Equatable {
-    case consentRequired
-    case ready
-}
-
-func hasAIChatExternalProviderConsent(userDefaults: UserDefaults) -> Bool {
-    userDefaults.bool(forKey: aiChatExternalProviderConsentUserDefaultsKey)
-}
-
-func grantAIChatExternalProviderConsent(userDefaults: UserDefaults) {
-    userDefaults.set(true, forKey: aiChatExternalProviderConsentUserDefaultsKey)
-}
-
-func aiChatAccessState(
-    hasExternalProviderConsent: Bool
-) -> AIChatAccessState {
-    guard hasExternalProviderConsent else {
-        return .consentRequired
-    }
-
-    return .ready
-}
-
-func isGuestAiLimitCode(_ code: String?) -> Bool {
-    code == "GUEST_AI_LIMIT_REACHED"
-}
-
 func aiChatAppVersion() -> String {
     appMarketingVersion()
 }
@@ -132,41 +61,6 @@ func aiChatTruncatedSnippet(_ value: String) -> String {
     return String(trimmed[..<endIndex]) + "..."
 }
 
-func aiChatCardAttachmentLabel(card: AIChatCardReference) -> String {
-    let snippet = aiChatTruncatedSnippet(card.frontText)
-    return snippet.isEmpty ? "Card" : "Card · \(snippet)"
-}
-
-func buildAIChatCardContextXML(card: AIChatCardReference) -> String {
-    let tagsXML = card.tags.map { tag in
-        "<tag>\(escapeAIChatCardXMLValue(tag))</tag>"
-    }.joined()
-
-    // Keep this byte-for-byte aligned with apps/backend/src/chat/cardContext.ts::buildCardContextXml.
-    return [
-        "<attached_card>",
-        "<card_id>\(escapeAIChatCardXMLValue(card.cardId))</card_id>",
-        "<effort_level>\(escapeAIChatCardXMLValue(card.effortLevel.rawValue))</effort_level>",
-        "<front_text>",
-        escapeAIChatCardXMLValue(card.frontText),
-        "</front_text>",
-        "<back_text>",
-        escapeAIChatCardXMLValue(card.backText),
-        "</back_text>",
-        "<tags>\(tagsXML)</tags>",
-        "</attached_card>",
-    ].joined(separator: "\n")
-}
-
-private func escapeAIChatCardXMLValue(_ value: String) -> String {
-    value
-        .replacingOccurrences(of: "&", with: "&amp;")
-        .replacingOccurrences(of: "<", with: "&lt;")
-        .replacingOccurrences(of: ">", with: "&gt;")
-        .replacingOccurrences(of: "\"", with: "&quot;")
-        .replacingOccurrences(of: "'", with: "&apos;")
-}
-
 func aiChatDecoderSummary(error: Error) -> String {
     if let decodingError = error as? DecodingError {
         switch decodingError {
@@ -211,58 +105,6 @@ private func aiChatStableSeedHash(_ value: String) -> String {
 
     return String(hash, radix: 16)
 }
-
-struct AIChatProviderDef: Codable, Hashable, Sendable {
-    let id: String
-    let label: String
-}
-
-struct AIChatReasoningDef: Codable, Hashable, Sendable {
-    let effort: String
-    let label: String
-}
-
-struct AIChatFeaturesDef: Codable, Hashable, Sendable {
-    let modelPickerEnabled: Bool
-    let dictationEnabled: Bool
-    let attachmentsEnabled: Bool
-}
-
-struct AIChatServerModelDef: Codable, Hashable, Sendable {
-    let id: String
-    let label: String
-    let badgeLabel: String
-}
-
-struct AIChatServerConfig: Codable, Hashable, Sendable {
-    let provider: AIChatProviderDef
-    let model: AIChatServerModelDef
-    let reasoning: AIChatReasoningDef
-    let features: AIChatFeaturesDef
-    let liveUrl: String?
-}
-
-let aiChatDefaultServerConfig = AIChatServerConfig(
-    provider: AIChatProviderDef(
-        id: "openai",
-        label: aiChatDefaultProviderLabel
-    ),
-    model: AIChatServerModelDef(
-        id: aiChatDefaultModelId,
-        label: aiChatDefaultModelLabel,
-        badgeLabel: "\(aiChatDefaultModelLabel) · \(aiChatDefaultReasoningLabel)"
-    ),
-    reasoning: AIChatReasoningDef(
-        effort: aiChatDefaultReasoningEffort,
-        label: aiChatDefaultReasoningLabel
-    ),
-    features: AIChatFeaturesDef(
-        modelPickerEnabled: false,
-        dictationEnabled: true,
-        attachmentsEnabled: true
-    ),
-    liveUrl: nil
-)
 
 enum AIChatRole: String, Codable, Hashable, Sendable {
     case user
@@ -629,7 +471,7 @@ private struct AIChatDecodableContentPartPayload: Decodable {
     }
 }
 
-private func decodeAIChatContentParts<Key: CodingKey>(
+func decodeAIChatContentParts<Key: CodingKey>(
     container: KeyedDecodingContainer<Key>,
     key: Key
 ) throws -> [AIChatContentPart] {
@@ -741,472 +583,7 @@ struct AIChatPersistedState: Codable, Hashable, Sendable {
     }
 }
 
-struct AIChatStartRunRequestBody: Codable, Hashable, Sendable {
-    let sessionId: String?
-    let clientRequestId: String
-    let content: [AIChatContentPart]
-    let timezone: String
-}
-
-enum AIChatComposerPhase: String, Hashable, Sendable {
-    case idle
-    case preparingSend
-    case startingRun
-    case running
-    case stopping
-}
-
-enum AIChatBootstrapPhase: Hashable, Sendable {
-    case ready
-    case loading
-    case failed(String)
-}
-
-struct AIChatConversation: Hashable, Sendable {
-    let messages: [AIChatMessage]
-    let updatedAt: Int
-    let mainContentInvalidationVersion: Int
-    let hasOlder: Bool
-    let oldestCursor: String?
-}
-
-struct AIChatActiveRunLive: Hashable, Sendable {
-    let cursor: String?
-    let stream: AIChatLiveStreamEnvelope
-}
-
-struct AIChatActiveRun: Hashable, Sendable {
-    let runId: String
-    let status: String
-    let live: AIChatActiveRunLive
-    let lastHeartbeatAt: Int?
-}
-
-struct AIChatConversationEnvelope: Hashable, Sendable {
-    let sessionId: String
-    let conversationScopeId: String
-    let conversation: AIChatConversation
-    let composerSuggestions: [AIChatComposerSuggestion]
-    let chatConfig: AIChatServerConfig
-    let activeRun: AIChatActiveRun?
-}
-
-typealias AIChatSessionSnapshot = AIChatConversationEnvelope
-typealias AIChatBootstrapResponse = AIChatConversationEnvelope
-
-struct AIChatConversationPayload: Decodable, Hashable, Sendable {
-    let messages: [AIChatConversationMessagePayload]
-    let updatedAt: Int
-    let mainContentInvalidationVersion: Int
-    let hasOlder: Bool?
-    let oldestCursor: String?
-}
-
-struct AIChatActiveRunLivePayload: Decodable, Hashable, Sendable {
-    let cursor: String?
-    let stream: AIChatLiveStreamEnvelope
-}
-
-struct AIChatActiveRunPayload: Decodable, Hashable, Sendable {
-    let runId: String
-    let status: String
-    let live: AIChatActiveRunLivePayload
-    let lastHeartbeatAt: Int?
-
-    private enum CodingKeys: String, CodingKey {
-        case runId
-        case status
-        case live
-        case lastHeartbeatAt
-        case lastHeartbeatAtMillis
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.runId = try container.decode(String.self, forKey: .runId)
-        self.status = try container.decode(String.self, forKey: .status)
-        self.live = try container.decode(AIChatActiveRunLivePayload.self, forKey: .live)
-        self.lastHeartbeatAt = try container.decodeIfPresent(Int.self, forKey: .lastHeartbeatAt)
-            ?? container.decodeIfPresent(Int.self, forKey: .lastHeartbeatAtMillis)
-    }
-}
-
-struct AIChatConversationEnvelopePayload: Decodable, Hashable, Sendable {
-    let sessionId: String
-    let conversationScopeId: String
-    let conversation: AIChatConversationPayload
-    let composerSuggestions: [AIChatComposerSuggestion]
-    let chatConfig: AIChatServerConfig
-    let activeRun: AIChatActiveRunPayload?
-
-    private enum CodingKeys: String, CodingKey {
-        case sessionId
-        case conversationScopeId
-        case conversation
-        case composerSuggestions
-        case chatConfig
-        case activeRun
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.sessionId = try container.decode(String.self, forKey: .sessionId)
-        self.conversationScopeId = try container.decode(String.self, forKey: .conversationScopeId)
-        self.conversation = try container.decode(AIChatConversationPayload.self, forKey: .conversation)
-        self.composerSuggestions = try container.decodeIfPresent(
-            [AIChatComposerSuggestion].self,
-            forKey: .composerSuggestions
-        ) ?? []
-        self.chatConfig = try container.decode(AIChatServerConfig.self, forKey: .chatConfig)
-        self.activeRun = try container.decodeIfPresent(AIChatActiveRunPayload.self, forKey: .activeRun)
-    }
-}
-
-typealias AIChatSessionSnapshotPayload = AIChatConversationEnvelopePayload
-typealias AIChatBootstrapResponsePayload = AIChatConversationEnvelopePayload
-
-struct AIChatConversationMessagePayload: Decodable, Hashable, Sendable {
-    let role: AIChatRole
-    let content: [AIChatContentPart]
-    let timestamp: String
-    let isError: Bool
-    let isStopped: Bool
-    let cursor: String?
-    let itemId: String?
-
-    private enum CodingKeys: String, CodingKey {
-        case role
-        case content
-        case timestamp
-        case timestampMillis
-        case isError
-        case isStopped
-        case cursor
-        case itemId
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.role = try container.decode(AIChatRole.self, forKey: .role)
-        self.content = try decodeAIChatContentParts(container: container, key: .content)
-        if let timestampMillis = try? container.decode(Int.self, forKey: .timestamp) {
-            self.timestamp = isoTimestampFromMilliseconds(timestampMillis)
-        } else if let timestamp = try? container.decode(String.self, forKey: .timestamp) {
-            self.timestamp = timestamp
-        } else if let timestampMillis = try? container.decode(Int.self, forKey: .timestampMillis) {
-            self.timestamp = isoTimestampFromMilliseconds(timestampMillis)
-        } else {
-            throw DecodingError.keyNotFound(
-                CodingKeys.timestamp,
-                DecodingError.Context(
-                    codingPath: decoder.codingPath,
-                    debugDescription: "AI chat message timestamp is missing."
-                )
-            )
-        }
-        self.isError = try container.decode(Bool.self, forKey: .isError)
-        self.isStopped = try container.decodeIfPresent(Bool.self, forKey: .isStopped) ?? false
-        self.cursor = try container.decodeIfPresent(String.self, forKey: .cursor)
-        self.itemId = try container.decodeIfPresent(String.self, forKey: .itemId)
-    }
-}
-
-// MARK: - Thin Live Chat Types
-
-struct AIChatOlderMessagesResponse: Sendable {
-    let messages: [AIChatMessage]
-    let hasOlder: Bool
-    let oldestCursor: String?
-}
-
-enum AIChatRunTerminalOutcome: String, Decodable, Hashable, Sendable {
-    case completed
-    case stopped
-    case error
-    case resetRequired = "reset_required"
-}
-
-struct AIChatLiveEventMetadata: Hashable, Sendable {
-    let sessionId: String
-    let conversationScopeId: String
-    let runId: String
-    let cursor: String?
-    let sequenceNumber: Int
-    let streamEpoch: String
-}
-
-struct AIChatComposerSuggestion: Codable, Hashable, Sendable, Identifiable {
-    let id: String
-    let text: String
-    let source: String
-    let assistantItemId: String?
-}
-
-enum AIChatLiveEvent: Sendable {
-    case assistantDelta(metadata: AIChatLiveEventMetadata, text: String, itemId: String)
-    case assistantToolCall(metadata: AIChatLiveEventMetadata, toolCall: AIChatToolCall, itemId: String)
-    case assistantReasoningStarted(metadata: AIChatLiveEventMetadata, reasoningId: String, itemId: String)
-    case assistantReasoningSummary(metadata: AIChatLiveEventMetadata, reasoningId: String, summary: String, itemId: String)
-    case assistantReasoningDone(metadata: AIChatLiveEventMetadata, reasoningId: String, itemId: String)
-    case assistantMessageDone(
-        metadata: AIChatLiveEventMetadata,
-        itemId: String,
-        content: [AIChatContentPart],
-        isError: Bool,
-        isStopped: Bool
-    )
-    case composerSuggestionsUpdated(
-        metadata: AIChatLiveEventMetadata,
-        suggestions: [AIChatComposerSuggestion]
-    )
-    case repairStatus(metadata: AIChatLiveEventMetadata, status: AIChatRepairAttemptStatus)
-    case runTerminal(
-        metadata: AIChatLiveEventMetadata,
-        outcome: AIChatRunTerminalOutcome,
-        message: String?,
-        assistantItemId: String?,
-        isError: Bool?,
-        isStopped: Bool?
-    )
-}
-
-struct AIChatAcceptedConversationEnvelopePayload: Decodable, Hashable, Sendable {
-    let accepted: Bool
-    let sessionId: String
-    let conversationScopeId: String
-    let conversation: AIChatConversationPayload
-    let composerSuggestions: [AIChatComposerSuggestion]
-    let chatConfig: AIChatServerConfig
-    let activeRun: AIChatActiveRunPayload?
-    let deduplicated: Bool?
-
-    private enum CodingKeys: String, CodingKey {
-        case accepted
-        case sessionId
-        case conversationScopeId
-        case conversation
-        case composerSuggestions
-        case chatConfig
-        case activeRun
-        case deduplicated
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.accepted = try container.decode(Bool.self, forKey: .accepted)
-        self.sessionId = try container.decode(String.self, forKey: .sessionId)
-        self.conversationScopeId = try container.decode(String.self, forKey: .conversationScopeId)
-        self.conversation = try container.decode(AIChatConversationPayload.self, forKey: .conversation)
-        self.composerSuggestions = try container.decodeIfPresent(
-            [AIChatComposerSuggestion].self,
-            forKey: .composerSuggestions
-        ) ?? []
-        self.chatConfig = try container.decode(AIChatServerConfig.self, forKey: .chatConfig)
-        self.activeRun = try container.decodeIfPresent(AIChatActiveRunPayload.self, forKey: .activeRun)
-        self.deduplicated = try container.decodeIfPresent(Bool.self, forKey: .deduplicated)
-    }
-}
-
-struct AIChatStartRunResponse: Hashable, Sendable {
-    let accepted: Bool
-    let sessionId: String
-    let conversationScopeId: String
-    let conversation: AIChatConversation
-    let composerSuggestions: [AIChatComposerSuggestion]
-    let chatConfig: AIChatServerConfig
-    let activeRun: AIChatActiveRun?
-    let deduplicated: Bool?
-
-    var envelope: AIChatConversationEnvelope {
-        AIChatConversationEnvelope(
-            sessionId: self.sessionId,
-            conversationScopeId: self.conversationScopeId,
-            conversation: self.conversation,
-            composerSuggestions: self.composerSuggestions,
-            chatConfig: self.chatConfig,
-            activeRun: self.activeRun
-        )
-    }
-}
-
-struct AIChatLiveStreamEnvelope: Codable, Hashable, Sendable {
-    let url: String
-    let authorization: String
-    let expiresAt: Int
-}
-
-struct AIChatNewSessionRequestBody: Codable, Hashable, Sendable {
-    let sessionId: String?
-}
-
-struct AIChatNewSessionResponse: Codable, Hashable, Sendable {
-    let ok: Bool
-    let sessionId: String
-    let composerSuggestions: [AIChatComposerSuggestion]
-    let chatConfig: AIChatServerConfig
-
-    private enum CodingKeys: String, CodingKey {
-        case ok
-        case sessionId
-        case composerSuggestions
-        case chatConfig
-    }
-
-    init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.ok = try container.decode(Bool.self, forKey: .ok)
-        self.sessionId = try container.decode(String.self, forKey: .sessionId)
-        self.composerSuggestions = try container.decodeIfPresent(
-            [AIChatComposerSuggestion].self,
-            forKey: .composerSuggestions
-        ) ?? []
-        self.chatConfig = try container.decode(AIChatServerConfig.self, forKey: .chatConfig)
-    }
-}
-
-struct AIChatStopRunResponse: Decodable, Hashable, Sendable {
-    let sessionId: String
-    let conversationScopeId: String
-    let runId: String?
-    let stopped: Bool
-    let stillRunning: Bool
-}
-
-struct AIToolCallRequest: Hashable, Sendable {
-    let toolCallId: String
-    let name: String
-    let input: String
-}
-
-enum AIChatFailureStage: String, Codable, Hashable, Sendable {
-    case requestBuild = "request_build"
-    case invalidHttpResponse = "invalid_http_response"
-    case httpResponseBody = "http_response_body"
-    case responseNotOk = "response_not_ok"
-    case readingLine = "reading_line"
-    case finishingEvent = "finishing_event"
-    case decodingEventJSON = "decoding_event_json"
-    case processingTrailingEvent = "processing_trailing_event"
-    case backendErrorEvent = "backend_error_event"
-    case toolInputDecode = "tool_input_decode"
-    case toolExecution = "tool_execution"
-}
-
-enum AIChatFailureKind: String, Codable, Hashable, Sendable {
-    case invalidBaseUrl = "invalid_base_url"
-    case invalidStreamResponse = "invalid_stream_response"
-    case invalidHttpResponse = "invalid_http_response"
-    case invalidSSEFraming = "invalid_sse_framing"
-    case invalidSSEEventJSON = "invalid_sse_event_json"
-    case invalidStreamContract = "invalid_stream_contract"
-    case backendErrorEvent = "backend_error_event"
-    case invalidToolInput = "invalid_tool_input"
-    case toolExecutionFailed = "tool_execution_failed"
-}
-
-struct AIChatFailureDiagnostics: Codable, Hashable, Sendable {
-    let clientRequestId: String
-    let backendRequestId: String?
-    let stage: AIChatFailureStage
-    let errorKind: AIChatFailureKind
-    let statusCode: Int?
-    let eventType: String?
-    let toolName: String?
-    let toolCallId: String?
-    let lineNumber: Int?
-    let rawSnippet: String?
-    let decoderSummary: String?
-    let continuationAttempt: Int?
-    let continuationToolCallIds: [String]
-}
-
-protocol AIChatFailureDiagnosticProviding: Error {
-    var diagnostics: AIChatFailureDiagnostics { get }
-}
-
-struct AIChatRepairAttemptStatus: Hashable, Sendable {
-    let message: String
-    let attempt: Int
-    let maxAttempts: Int
-    let toolName: String?
-
-    var displayText: String {
-        "\(self.message) \(self.attempt)/\(self.maxAttempts)"
-    }
-}
-
-enum AIChatRuntimeEvent: Sendable {
-    case accepted(AIChatStartRunResponse)
-    case liveEvent(AIChatLiveEvent)
-    case appendAssistantAccountUpgradePrompt(message: String, buttonTitle: String)
-    case finish
-    case fail(String)
-}
-
-struct AIChatContext: Sendable {
-    let workspace: Workspace
-    let schedulerSettings: WorkspaceSchedulerSettings
-    let totalActiveCards: Int
-}
-
-struct AIChatResumeAttemptDiagnostics: Equatable, Sendable {
-    let sequence: Int
-
-    var headerValue: String {
-        String(self.sequence)
-    }
-}
-
-protocol AIChatHistoryStoring: Sendable {
-    func activateWorkspace(workspaceId: String?)
-    func loadState() -> AIChatPersistedState
-    func saveState(state: AIChatPersistedState) async
-    func clearState() async
-    func loadDraft(workspaceId: String?, sessionId: String?) -> AIChatComposerDraft
-    func saveDraft(workspaceId: String?, sessionId: String?, draft: AIChatComposerDraft) async
-}
-
-protocol AIChatSessionServicing: Sendable {
-    func loadSnapshot(
-        session: CloudLinkedSession,
-        sessionId: String?
-    ) async throws -> AIChatSessionSnapshot
-
-    func loadBootstrap(
-        session: CloudLinkedSession,
-        sessionId: String?,
-        limit: Int,
-        resumeAttemptDiagnostics: AIChatResumeAttemptDiagnostics?
-    ) async throws -> AIChatBootstrapResponse
-
-    func loadOlderMessages(
-        session: CloudLinkedSession,
-        sessionId: String,
-        beforeCursor: String,
-        limit: Int
-    ) async throws -> AIChatOlderMessagesResponse
-
-    func startRun(
-        session: CloudLinkedSession,
-        request: AIChatStartRunRequestBody
-    ) async throws -> AIChatStartRunResponse
-
-    func createNewSession(
-        session: CloudLinkedSession,
-        sessionId: String?
-    ) async throws -> AIChatNewSessionResponse
-
-    func stopRun(
-        session: CloudLinkedSession,
-        sessionId: String
-    ) async throws -> AIChatStopRunResponse
-}
-
-protocol AIChatContextLoading: Sendable {
-    func loadContext() async throws -> AIChatContext
-}
-
-private func isoTimestampFromMilliseconds(_ value: Int) -> String {
+func isoTimestampFromMilliseconds(_ value: Int) -> String {
     let date = Date(timeIntervalSince1970: TimeInterval(value) / 1000)
     return isoTimestampFromDate(date)
 }
