+++ apps/ios/Flashcards/Flashcards/AI/AIChatService.swift
@@ -77,7 +77,7 @@ final class AIChatService: AIChatSessionServicing, @unchecked Sendable {
         )
 
         do {
-            let payload = try self.decoder.decode(AIChatSessionSnapshotPayload.self, from: data)
+            let payload = try self.decoder.decode(AIChatSessionSnapshotWire.self, from: data)
             return mapConversationEnvelope(payload)
         } catch {
             let diagnostics = AIChatFailureDiagnostics(
@@ -126,7 +126,7 @@ final class AIChatService: AIChatSessionServicing, @unchecked Sendable {
         )
 
         do {
-            let payload = try self.decoder.decode(AIChatBootstrapResponsePayload.self, from: data)
+            let payload = try self.decoder.decode(AIChatBootstrapResponseWire.self, from: data)
             return mapConversationEnvelope(payload)
         } catch {
             let diagnostics = AIChatFailureDiagnostics(
@@ -173,7 +173,7 @@ final class AIChatService: AIChatSessionServicing, @unchecked Sendable {
         )
 
         do {
-            let payload = try self.decoder.decode(AIChatBootstrapResponsePayload.self, from: data)
+            let payload = try self.decoder.decode(AIChatBootstrapResponseWire.self, from: data)
             return AIChatOlderMessagesResponse(
                 messages: payload.conversation.messages.enumerated().map { index, message in
                     mapConversationMessage(
@@ -224,7 +224,7 @@ final class AIChatService: AIChatSessionServicing, @unchecked Sendable {
         )
 
         do {
-            let payload = try self.decoder.decode(AIChatAcceptedConversationEnvelopePayload.self, from: data)
+            let payload = try self.decoder.decode(AIChatAcceptedConversationEnvelopeWire.self, from: data)
             return mapAcceptedConversationEnvelope(payload)
         } catch {
             let diagnostics = AIChatFailureDiagnostics(
@@ -463,88 +463,6 @@ final class AIChatService: AIChatSessionServicing, @unchecked Sendable {
 
         return url
     }
-}
-
-private func mapConversationEnvelope(
-    _ payload: AIChatConversationEnvelopePayload
-) -> AIChatConversationEnvelope {
-    AIChatConversationEnvelope(
-        sessionId: payload.sessionId,
-        conversationScopeId: payload.conversationScopeId,
-        conversation: AIChatConversation(
-            messages: payload.conversation.messages.enumerated().map { index, message in
-                mapConversationMessage(
-                    sessionId: payload.sessionId,
-                    index: index,
-                    message: message
-                )
-            },
-            updatedAt: payload.conversation.updatedAt,
-            mainContentInvalidationVersion: payload.conversation.mainContentInvalidationVersion,
-            hasOlder: payload.conversation.hasOlder ?? false,
-            oldestCursor: payload.conversation.oldestCursor
-        ),
-        composerSuggestions: payload.composerSuggestions,
-        chatConfig: payload.chatConfig,
-        activeRun: payload.activeRun.map(mapActiveRun)
-    )
-}
-
-private func mapAcceptedConversationEnvelope(
-    _ payload: AIChatAcceptedConversationEnvelopePayload
-) -> AIChatStartRunResponse {
-    AIChatStartRunResponse(
-        accepted: payload.accepted,
-        sessionId: payload.sessionId,
-        conversationScopeId: payload.conversationScopeId,
-        conversation: AIChatConversation(
-            messages: payload.conversation.messages.enumerated().map { index, message in
-                mapConversationMessage(
-                    sessionId: payload.sessionId,
-                    index: index,
-                    message: message
-                )
-            },
-            updatedAt: payload.conversation.updatedAt,
-            mainContentInvalidationVersion: payload.conversation.mainContentInvalidationVersion,
-            hasOlder: payload.conversation.hasOlder ?? false,
-            oldestCursor: payload.conversation.oldestCursor
-        ),
-        composerSuggestions: payload.composerSuggestions,
-        chatConfig: payload.chatConfig,
-        activeRun: payload.activeRun.map(mapActiveRun),
-        deduplicated: payload.deduplicated
-    )
-}
-
-private func mapConversationMessage(
-    sessionId: String,
-    index: Int,
-    message: AIChatConversationMessagePayload
-) -> AIChatMessage {
-    let stableCursor = message.cursor ?? "snapshot-\(index)"
-    return AIChatMessage(
-        id: "\(sessionId)-\(index)-\(stableCursor)",
-        role: message.role,
-        content: message.content,
-        timestamp: message.timestamp,
-        isError: message.isError,
-        isStopped: message.isStopped,
-        cursor: message.cursor,
-        itemId: message.itemId
-    )
-}
-
-private func mapActiveRun(_ payload: AIChatActiveRunPayload) -> AIChatActiveRun {
-    AIChatActiveRun(
-        runId: payload.runId,
-        status: payload.status,
-        live: AIChatActiveRunLive(
-            cursor: payload.live.cursor,
-            stream: payload.live.stream
-        ),
-        lastHeartbeatAt: payload.lastHeartbeatAt
-    )
 }
 
 private func extractChatRequestId(httpResponse: HTTPURLResponse) -> String? {
